### 2025-02-20 #7801 @editorialbot
[PRE REVIEW]: SeqIKPy: a Python package for inverse kinematics in insects
**Submitting author:** @gizemozd (<a href="http://orcid.org/0000-0003-4507-6642">Pembe Gizem Ozdil</a>)
**Repository:** https://github.com/NeLy-EPFL/sequential-inverse-kinematics
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.3
