+++ docs/paper/paper.bib
@@ -15,14 +15,14 @@
 }
 
 @software{Manceron_IKPy,
-  author  = {Manceron, Pierre},
-  doi     = {10.5281/zenodo.6551105},
-  license = {GPL-2.0},
-  title   = {{IKPy}},
-  year = {2016},
+  author    = {Manceron, Pierre},
+  doi       = {10.5281/zenodo.6551105},
+  license   = {GPL-2.0},
+  title     = {{IKPy}},
+  year      = {2016},
   publisher = {GitHub},
-  journal = {GitHub repository},
-  url     = {https://github.com/Phylliade/ikpy}
+  journal   = {GitHub repository},
+  url       = {https://github.com/Phylliade/ikpy}
 }
 
 
@@ -61,6 +61,7 @@
   author    = {Pereira, Talmo D and Shaevitz, Joshua W and Murthy, Mala},
   journal   = {Nature neuroscience},
   volume    = {23},
+  doi       = {10.1038/s41593-020-00734-z},
   number    = {12},
   pages     = {1537--1549},
   year      = {2020},
@@ -73,6 +74,7 @@
   journal   = {Journal of biomechanical engineering},
   volume    = {140},
   number    = {3},
+  doi       = {10.1115/1.4038741},
   pages     = {030801},
   year      = {2018},
   publisher = {American Society of Mechanical Engineers}
@@ -84,6 +86,7 @@
   journal   = {IEEE transactions on biomedical engineering},
   volume    = {54},
   number    = {11},
+  doi       = {10.1109/TBME.2007.901024},
   pages     = {1940--1950},
   year      = {2007},
   publisher = {IEEE}
@@ -96,6 +99,7 @@
   volume       = {37},
   number       = {6},
   pages        = {35--58},
+  doi          = {10.1111/cgf.13310},
   year         = {2018},
   organization = {Wiley Online Library}
 }
@@ -107,6 +111,7 @@
   journal   = {Plos one},
   volume    = {18},
   number    = {11},
+  doi       = {10.1371/journal.pone.0295152},
   pages     = {e0295152},
   year      = {2023},
   publisher = {Public Library of Science San Francisco, CA USA}
@@ -120,7 +125,6 @@
   urldate   = {2024-03-15},
   publisher = {bioRxiv},
   author    = {Vaxenburg, Roman and Siwanowicz, Igor and Merel, Josh and Robie, Alice A. and Morrow, Carmen and Novati, Guido and Stefanidi, Zinovia and Card, Gwyneth M. and Reiser, Michael B. and Botvinick, Matthew M. and Branson, Kristin M. and Tassa, Yuval and Turaga, Srinivas C.},
-  month     = mar,
   year      = {2024}
 }
 
@@ -129,13 +133,13 @@
   author    = {Wang-Chen, Sibo and Stimpfling, Victor Alfred and Lam, Thomas Ka Chung and {\"O}zdil, Pembe Gizem and Genoud, Louise and Hurtak, Femke and Ramdya, Pavan},
   journal   = {Nature Methods},
   pages     = {1--10},
+  doi       = {10.1038/s41592-024-02497-y},
   year      = {2024},
   publisher = {Nature Publishing Group US New York}
 }
 
 @article{pagnon:2022,
   doi       = {10.21105/joss.04362},
-  url       = {https://doi.org/10.21105/joss.04362},
   year      = {2022},
   publisher = {The Open Journal},
   volume    = {7},
@@ -151,6 +155,7 @@
   author       = {Todorov, Emanuel and Erez, Tom and Tassa, Yuval},
   booktitle    = {2012 IEEE/RSJ international conference on intelligent robots and systems},
   pages        = {5026--5033},
+  doi          = {10.1109/IROS.2012.6386109},
   year         = {2012},
   organization = {IEEE}
 }
