+++ README.md
@@ -7,6 +7,7 @@
 
 [![License](https://img.shields.io/badge/License-Apache_2.0-red)](https://opensource.org/license/apache-2-0)
 [![python](https://img.shields.io/badge/python-3.9%2C%203.10-blue)]()
+[![Cov](https://github.com/NeLy-EPFL/sequential-inverse-kinematics/blob/main/docs/images/coverage.svg?raw=true)]()
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.12601317.svg)](https://doi.org/10.5281/zenodo.12601317)
 </div>
 
