### 2025-02-21 #7801 @gizemozd
Hi @Kevin-Mattheus-Moerman,

- _Please study the above reference check ☝️ and see if you can address any of the reported potential DOI issues. You can add/amend DOI entries in your .bib file, and call @editorialbot check references here to check them again._
I have addressed the missing DOIs and confirmed it with editorialbot, please see the comments above.

- _Can you confirm this project features automated testing? If so can you clearly link to this in the README (or add a badge related to it)?_
We have a GitHub workflow for automated testing, see [here](https://github.com/NeLy-EPFL/sequential-inverse-kinematics/blob/1598cf49a166f061695c01b5feace398be2a5cac/.github/workflows/CI.yml). Additionally, I have added a badge for code coverage in [README](https://github.com/NeLy-EPFL/sequential-inverse-kinematics/blob/1598cf49a166f061695c01b5feace398be2a5cac/README.md).

- _Minor side note (not a requirement) your logo looks like the below for folks with a dark GitHub theme, hence you may want to consider adding a white boundary or background to the insect bit:_
I have addressed this in the new [README](https://github.com/NeLy-EPFL/sequential-inverse-kinematics/blob/1598cf49a166f061695c01b5feace398be2a5cac/README.md) as well.

Please let me know if anything is still missing. Thank you very much for handling our submission!
