### 2025-04-16 #7801 @olexandr-konovalov
Hi @davidpagnon and @melund - I am looking for reviewers for this JOSS paper, and your names were suggested in the thread above as potential reviewers. Would you be able to review this submission for JOSS, please?

Just in case if you're not familiar with the JOSS publishing model, JOSS reviews take place in GitHub issues and are public, with author-reviewer-editor conversations being encouraged. Reviewing involves downloading and installing the software, and inspecting the repository and submitted paper for key elements (see https://joss.readthedocs.io/en/latest/review_criteria.html for further details). Editors and reviewers post comments on the Review issue, and authors respond to the comments and improve their submission until acceptance (or withdrawal, if they feel unable to satisfy the review). You can find more information in the reviewer's guide at https://joss.readthedocs.io/en/latest/, and also see some other review issues in this repository for actual examples of JOSS reviews (in particular, reviewer's checklists).

If not, could you possibly recommend someone whom we may approach regarding this?
