### 2025-04-16 #7801 @davidpagnon
Why not! I'm quite busy and I won't be able to work on it this week, but I'll try to dig into it next week.
