### 2025-04-17 #7801 @olexandr-konovalov
Hi @trallard and @tuliofalmeida - I am looking for the 2nd reviewer for this JOSS paper, and your names were suggested in the thread above as potential reviewers. Would one of you be able to review this submission for JOSS, please?

@melund - thanks for the quick reply!
