### 2025-07-07 #7801 @aceglia
Hi @olexandr-konovalov, I agree to review this JOSS paper. I'm waiting for more information on how to do so. Amedeo
