### 2025-07-09 #7801 @olexandr-konovalov
Many thanks again @aceglia and @davidpagnon for agreeing to review this paper! Further conversation moves to #8557 - please see it for further instructions, generating your review checklists etc.
