### 2025-07-09 #7801 @heath-henninger
Please remove me from this chain.  Thanks.

Heath

From: The Open Journals editorial robot ***@***.***>
Sent: Wednesday, July 9, 2025 3:02 AM
Cc: Heath Byron Henninger ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: SeqIKPy: a Python package for inverse kinematics in insects (Issue #7801)

[https://avatars.githubusercontent.com/u/89919391?s=20&v=4]editorialbot left a comment (openjournals/joss-reviews#7801)<https://github.com/openjournals/joss-reviews/issues/7801#issuecomment-3051784070>

OK, I've started the review over in #8557<https://github.com/openjournals/joss-reviews/issues/8557>.
