### 2025-10-01 #8557 @olexandr-konovalov
Many thanks @aceglia and @davidpagnon for the detailed reviews! Sorry, over the summer (still with some teaching) and start of the autumn semester, it took a while to acknowledge it. Very useful comments, leading not only to the improvement of the paper, but the software itself, e.g. documentation & tutorials.

@gizemozd could you reply to the reviewers' suggestions please? Which are critical for the acceptance, which are easy to implement and should be done before acceptance, which needs further discussions, e.g. the use of Kalman filters, etc.? Will you be making changes in the code//documentation/tutorials? Would you like to ask reviewers to convert suggestions into issues in your repository, like we recommend at [reviewing guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)?
