### 2025-10-13 6fc94e7
bump up python and numpy versions
