### 2025-10-13 e505745
Parallelize at leg level
