+++ seqikpy/leg_inverse_kinematics.py
@@ -8,6 +8,7 @@ import warnings
 
 import numpy as np
 from tqdm import trange
+from joblib import Parallel, delayed
 from ikpy.chain import Chain
 
 from seqikpy.utils import save_file
@@ -207,6 +208,43 @@ class LegInvKinSeq(LegInvKinBase):
         # Create an empty dict for joint angles
         self.joint_angles_dict = {}
 
+    @staticmethod
+    def _process_single_leg(segment_name, segment_array, stages, kinematic_chain_class, initial_angles_dict, hide_progress_bar):
+        """Process a single leg through all stages."""
+        leg_name = segment_name.split("_")[0]
+        
+        if not leg_name in kinematic_chain_class.body_size:
+            return segment_name, None, {}
+        
+        # Create a temporary instance to process this leg
+        temp_instance = LegInvKinSeq.__new__(LegInvKinSeq)
+        temp_instance.kinematic_chain_class = kinematic_chain_class
+        temp_instance.initial_angles = initial_angles_dict
+        temp_instance.joint_angles_dict = {}
+        
+        # Setup minimal logger
+        import logging
+        temp_instance.logger = logging.getLogger("LegInvKinSeq")
+        temp_instance.logger.setLevel(logging.INFO)
+        
+        origin = segment_array[:, 0, :]
+        forward_kinematics_result = None
+        
+        for stage in stages:
+            end_effector_pos = segment_array[:, stage, :]
+            initial_angles = initial_angles_dict[leg_name][f"stage_{stage}"]
+            
+            forward_kinematics_result = temp_instance.calculate_ik_stage(
+                end_effector_pos=end_effector_pos,
+                origin=origin,
+                initial_angles=initial_angles,
+                stage=stage,
+                segment_name=leg_name,
+                hide_progress_bar=hide_progress_bar,
+            )
+        
+        return segment_name, forward_kinematics_result, temp_instance.joint_angles_dict
+
     def calculate_ik_stage(
         self,
         end_effector_pos: np.ndarray,
@@ -337,7 +375,11 @@ class LegInvKinSeq(LegInvKinBase):
         return forward_kinematics
 
     def run_ik_and_fk(
-        self, export_path: Union[Path, str] = None, **kwargs
+        self,
+        export_path: Union[Path, str] = None,
+        n_workers: int = 1,
+        hide_progress_bar: bool = False,
+        **kwargs
     ) -> Tuple[Dict[str, np.ndarray], Dict[str, np.ndarray]]:
         """Runs inverse and forward kinematics for leg joints.
 
@@ -348,6 +390,9 @@ class LegInvKinSeq(LegInvKinBase):
             if None, nothing is saveed, by default None
         stages (kwargs) : List[int], optional
             Stages to run the inverse kinematics.
+        n_workers (kwargs) : int, optional
+            Number of parallel jobs for leg processing. -1 uses all cores,
+            1 disables  parallelization, by default 1.
         hide_progress_bar (kwargs) : Optional[bool], optional
             Hide the progress bar, by default True
 
@@ -358,7 +403,6 @@ class LegInvKinSeq(LegInvKinBase):
             kinematics, respectively.
         """
         stages = kwargs.get("stages", [1, 2, 3, 4])
-        hide_progress_bar = kwargs.get("hide_progress_bar", False)
 
         if max(stages) > 4 or not all(np.diff(stages) == 1):
             raise ValueError(
@@ -366,42 +410,38 @@ class LegInvKinSeq(LegInvKinBase):
             )
         forward_kinematics_dict = {}
 
+        # Get parallel processing parameters        
         self.logger.info("Computing joint angles and forward kinematics...")
-        for segment_name, segment_array in self.aligned_pos.items():
[…]
