### 2025-10-13 9a6b596
add AI-generated tests
