+++ tests/test_parallelization.py
@@ -1,0 +1,334 @@
+"""Test parallelization functionality of inverse kinematics."""
+
+import pytest
+import numpy as np
+import time
+from pathlib import Path
+
+import seqikpy
+from seqikpy.leg_inverse_kinematics import LegInvKinSeq, LegInvKinGeneric
+from seqikpy.kinematic_chain import KinematicChainSeq, KinematicChainGeneric
+from seqikpy.data import BOUNDS, INITIAL_ANGLES
+from seqikpy.utils import calculate_body_size
+from seqikpy.data import NMF_TEMPLATE
+
+PKG_PATH = Path(seqikpy.__path__[0]).parent
+
+
+@pytest.fixture
+def synthetic_pose_data():
+    """Create synthetic test data for validation."""
+    n_frames = 50  # Smaller for faster tests
+    n_joints = 5
+    
+    aligned_pos = {}
+    leg_names = ["RF", "LF"]  # Test with 2 legs for speed
+    
+    # Set random seed for reproducible tests
+    np.random.seed(42)
+    
+    for leg_name in leg_names:
+        # Create synthetic 3D pose data for each leg
+        # Shape: (n_frames, n_joints, 3)
+        positions = np.random.randn(n_frames, n_joints, 3) * 0.1
+        
+        # Add some realistic structure - joints should be connected
+        for t in range(n_frames):
+            for joint in range(1, n_joints):
+                positions[t, joint] = positions[t, joint-1] + positions[t, joint]
+        
+        aligned_pos[f"{leg_name}_leg"] = positions
+    
+    return aligned_pos
+
+
+@pytest.fixture
+def test_body_size():
+    """Create body size for test legs."""
+    return calculate_body_size(NMF_TEMPLATE, ["RF", "LF"])
+
+
+class TestParallelization:
+    """Test class for parallelization functionality."""
+    
+    def test_sequential_ik_parallelization(self, synthetic_pose_data, test_body_size):
+        """Test that LegInvKinSeq gives identical results for sequential vs parallel processing."""
+        
+        kinematic_chain_seq = KinematicChainSeq(
+            bounds_dof=BOUNDS,
+            legs_list=["RF", "LF"],
+            body_size=test_body_size,
+        )
+        
+        # Sequential processing
+        seq_ik_sequential = LegInvKinSeq(
+            aligned_pos=synthetic_pose_data,
+            kinematic_chain_class=kinematic_chain_seq,
+            initial_angles=INITIAL_ANGLES,
+        )
+        
+        joint_angles_seq, fk_seq = seq_ik_sequential.run_ik_and_fk(
+            stages=[1, 2, 3, 4], hide_progress_bar=True, n_jobs=1
+        )
+        
+        # Parallel processing
+        seq_ik_parallel = LegInvKinSeq(
+            aligned_pos=synthetic_pose_data,
+            kinematic_chain_class=kinematic_chain_seq,
+            initial_angles=INITIAL_ANGLES,
+        )
+        
+        joint_angles_par, fk_par = seq_ik_parallel.run_ik_and_fk(
+            stages=[1, 2, 3, 4], hide_progress_bar=True, n_jobs=-1
+        )
+        
+        # Verify results are identical
+        assert len(joint_angles_seq) == len(joint_angles_par), "Different number of joint angles"
+        
+        for key in joint_angles_seq:
+            if key in joint_angles_par:
+                np.testing.assert_array_almost_equal(
+                    joint_angles_seq[key], 
+                    joint_angles_par[key],
+                    decimal=10,
+                    err_msg=f"Results differ for {key}"
+                )
+        
+        # Verify forward kinematics results are also identical
+        assert len(fk_seq) == len(fk_par), "Different number of forward kinematics results"
+        
+        for key in fk_seq:
+            if key in fk_par:
[…]
