### 2025-10-27 #8557 @olexandr-konovalov
Thank you for the reply @gizemozd, and all the best with the start at a new place! The delay is understandable - I hope you will find the time to finish what has been started and get the paper accepted. If necessary, please get in touch with the reviewers to clarify what is absolutely necessary for the acceptance, and what is a general suggestion to make the software better that can be addressed in the long run. I may send another reminder in a couple of months then!
