### 2025-11-21 0e85478
wip - implement parallelization over time
