+++ seqikpy/leg_inverse_kinematics.py
@@ -1,17 +1,18 @@
-""" Module to calculate inverse kinematics from 3D pose based on IKPy."""
+"""Module to calculate inverse kinematics from 3D pose based on IKPy."""
 
+import logging
+import warnings
+from collections import defaultdict
 from abc import ABC, abstractmethod
 from pathlib import Path
 from typing import Dict, Tuple, Union, Literal, Optional
-import logging
-import warnings
 
 import numpy as np
 from tqdm import trange
 from joblib import Parallel, delayed
 from ikpy.chain import Chain
 
-from seqikpy.utils import save_file
+from seqikpy.utils import save_file, split_arrays_into_chunks, merge_chunks_into_arrays
 from seqikpy.data import INITIAL_ANGLES
 from seqikpy.kinematic_chain import (
     KinematicChainBase,
@@ -22,10 +23,7 @@ from seqikpy.kinematic_chain import (
 # Ignore the warnings
 warnings.filterwarnings("ignore")
 
-logging.basicConfig(
-    format=" %(asctime)s - %(levelname)s- %(message)s",
-    handlers=[logging.StreamHandler()],
-)
+_logger = logging.getLogger(__name__)
 
 
 class LegInvKinBase(ABC):
@@ -63,9 +61,9 @@ class LegInvKinBase(ABC):
         )
 
         # Get the logger of the module
-        self.logger = logging.getLogger(self.__class__.__name__)
+        _logger = logging.getLogger(self.__class__.__name__)
         numeric_level = getattr(logging, log_level.upper(), None)
-        self.logger.setLevel(numeric_level)
+        _logger.setLevel(numeric_level)
 
     def calculate_ik(
         self,
@@ -209,31 +207,33 @@ class LegInvKinSeq(LegInvKinBase):
         self.joint_angles_dict = {}
 
     @staticmethod
-    def _process_single_leg(segment_name, segment_array, stages, kinematic_chain_class, initial_angles_dict, hide_progress_bar):
+    def _process_single_leg_sequence(
+        segment_name,
+        segment_array,
+        stages,
+        kinematic_chain_class,
+        initial_angles_dict,
+        hide_progress_bar,
+    ):
         """Process a single leg through all stages."""
         leg_name = segment_name.split("_")[0]
-        
+
         if not leg_name in kinematic_chain_class.body_size:
             return segment_name, None, {}
-        
+
         # Create a temporary instance to process this leg
         temp_instance = LegInvKinSeq.__new__(LegInvKinSeq)
         temp_instance.kinematic_chain_class = kinematic_chain_class
         temp_instance.initial_angles = initial_angles_dict
         temp_instance.joint_angles_dict = {}
-        
-        # Setup minimal logger
-        import logging
-        temp_instance.logger = logging.getLogger("LegInvKinSeq")
-        temp_instance.logger.setLevel(logging.INFO)
-        
+
         origin = segment_array[:, 0, :]
         forward_kinematics_result = None
-        
+
         for stage in stages:
             end_effector_pos = segment_array[:, stage, :]
             initial_angles = initial_angles_dict[leg_name][f"stage_{stage}"]
-            
+
             forward_kinematics_result = temp_instance.calculate_ik_stage(
                 end_effector_pos=end_effector_pos,
                 origin=origin,
@@ -242,7 +242,7 @@ class LegInvKinSeq(LegInvKinBase):
                 segment_name=leg_name,
                 hide_progress_bar=hide_progress_bar,
             )
-        
+
         return segment_name, forward_kinematics_result, temp_instance.joint_angles_dict
 
     def calculate_ik_stage(
[…]
