+++ seqikpy/utils.py
@@ -1,16 +1,14 @@
-""" Utilities. """
+"""Utilities."""
 
-from pathlib import Path
 import logging
-from typing import Dict, List
 import pickle
 import numpy as np
 import cv2
-from scipy.interpolate import pchip_interpolate
-
-import xml
 import xml.etree.ElementTree as ET
-import numpy as np
+from pathlib import Path
+from typing import Dict, List
+from collections import defaultdict
+from scipy.interpolate import pchip_interpolate
 
 
 def get_fps_from_video(video_dir):
@@ -431,3 +429,152 @@ def from_sdf(sdf_file: str):
         for joint in joints
     }
     return body_template, joint_bounds
+
+
+def split_arrays_into_chunks(
+    arrays: list[np.ndarray],
+    approx_n_chunks_total: int,
+    overlap: int = 100,
+    min_chunk_size: int = 1000,
+) -> list[tuple[int, int, np.ndarray]]:
+    """Splits arrays into chunks with overlap between chunks. This is useful for
+    processing long sequences from multiple kinematic chains in parallel.
+
+    Parameters
+    ----------
+    arrays : list[np.ndarray]
+        List of arrays to be split into chunks. Each array should have shape
+        (sequence_length, ...) where `...` can be any number of dimensions.
+    approx_n_chunks_total : int
+        Approximate number of chunks to split the arrays into. The actual number of
+        chunks is generally different in order to make the total number of chunks a
+        multiple of the number of arrays (i.e. kinematic chains).
+    overlap : int, optional
+        Number of overlapping frames between chunks, by default 100.
+    min_chunk_size : int, optional
+        Minimum size of each chunk, by default 1000.
+
+    Returns
+    -------
+    list[tuple[int, int, np.ndarray]]
+        List of tuples containing
+            - array_idx: int
+            - start_idx_within_array: int
+            - chunk_of_array: np.ndarray
+    """
+    assert overlap < min_chunk_size, "Overlap must be smaller than min_chunk_size."
+    n_arrays = len(arrays)
+    assert n_arrays > 0, "Number of arrays must be positive"
+    n_chunks_per_array = int(approx_n_chunks_total / n_arrays)
+    for arr in arrays:
+        assert arr.shape == arrays[0].shape, "All arrays must have the same shape."
+    assert arrays[0].ndim >= 1, "Arrays must be at least 1-dimensional."
+    seq_length = arrays[0].shape[0]
+
+    # If the chunk size is too small, return the entire arrays as single chunks
+    if seq_length <= min_chunk_size:
+        return [(arr_idx, 0, arr) for arr_idx, arr in enumerate(arrays)]
+
+    # Create chunks
+    chunk_size = seq_length // n_chunks_per_array  # last chunk is generally larger
+    if chunk_size < min_chunk_size:
+        n_chunks_per_array = seq_length // min_chunk_size
+        chunk_size = seq_length // n_chunks_per_array
+        logging.warning(
+            f"Requested number of chunks results in chunk size smaller than "
+            f"min_chunk_size. Reducing number of chunks per array to "
+            f"{n_chunks_per_array}."
+        )
+
+    chunks = []
+    for array_idx, arr in enumerate(arrays):
+        start_idx = 0
+        for chunk_idx in range(n_chunks_per_array):
+            # Determine the end index of the chunk before adding overlap
+            if chunk_idx == n_chunks_per_array - 1:
+                end_idx_before_overlap = seq_length
+                end_idx_after_overlap = seq_length
+            else:
+                end_idx_before_overlap = start_idx + chunk_size
+                end_idx_after_overlap = end_idx_before_overlap + overlap
+                assert end_idx_after_overlap <= seq_length, (
[…]

[…]
