### 2025-11-22 e7fa41f
get parallelization example working and run scaling test on SCITAS
