### 2025-11-22 2cdb390
refactor different processing modes into separate methods
