### 2025-11-22 0858467
fix typos in parallel pipeline
