+++ tests/test_parallelization.py
@@ -1,4 +1,23 @@
-"""Test parallelization functionality of inverse kinematics."""
+"""Test parallelization functionality of inverse kinematics.
+
+This test suite covers the parallelization features implemented in the parallel branch:
+
+1. Original parallelization over legs (using n_workers parameter, formerly n_jobs)
+2. New parallelization over both legs and time (parallel_over_time=True)
+3. Chunk processing utilities (split_arrays_into_chunks, merge_chunks_into_arrays)
+4. New parameters: chunk_overlap, avg_workloads_per_worker, min_chunk_size
+
+Key changes from original tests:
+- Updated parameter name from n_jobs to n_workers
+- Added tests for parallel_over_time functionality
+- Added comprehensive tests for chunk utilities
+- Updated tolerances for tests involving chunking (as chunking can introduce
+  boundary effects that result in slightly different optimization outcomes)
+
+Note: Parallel-over-time processing may produce slightly different results
+compared to sequential processing due to chunking artifacts at boundaries,
+especially for optimization-based inverse kinematics. This is expected behavior.
+"""
 
 import pytest
 import numpy as np
@@ -9,7 +28,11 @@ import seqikpy
 from seqikpy.leg_inverse_kinematics import LegInvKinSeq, LegInvKinGeneric
 from seqikpy.kinematic_chain import KinematicChainSeq, KinematicChainGeneric
 from seqikpy.data import BOUNDS, INITIAL_ANGLES
-from seqikpy.utils import calculate_body_size
+from seqikpy.utils import (
+    calculate_body_size,
+    split_arrays_into_chunks,
+    merge_chunks_into_arrays,
+)
 from seqikpy.data import NMF_TEMPLATE
 
 PKG_PATH = Path(seqikpy.__path__[0]).parent
@@ -18,27 +41,27 @@ PKG_PATH = Path(seqikpy.__path__[0]).parent
 @pytest.fixture
 def synthetic_pose_data():
     """Create synthetic test data for validation."""
-    n_frames = 50  # Smaller for faster tests
+    n_frames = 120  # Increased for better testing of chunking
     n_joints = 5
-    
+
     aligned_pos = {}
     leg_names = ["RF", "LF"]  # Test with 2 legs for speed
-    
+
     # Set random seed for reproducible tests
     np.random.seed(42)
-    
+
     for leg_name in leg_names:
         # Create synthetic 3D pose data for each leg
         # Shape: (n_frames, n_joints, 3)
         positions = np.random.randn(n_frames, n_joints, 3) * 0.1
-        
+
         # Add some realistic structure - joints should be connected
         for t in range(n_frames):
             for joint in range(1, n_joints):
-                positions[t, joint] = positions[t, joint-1] + positions[t, joint]
-        
+                positions[t, joint] = positions[t, joint - 1] + positions[t, joint]
+
         aligned_pos[f"{leg_name}_leg"] = positions
-    
+
     return aligned_pos
 
 
@@ -50,157 +73,168 @@ def test_body_size():
 
 class TestParallelization:
     """Test class for parallelization functionality."""
-    
+
     def test_sequential_ik_parallelization(self, synthetic_pose_data, test_body_size):
         """Test that LegInvKinSeq gives identical results for sequential vs parallel processing."""
-        
+
         kinematic_chain_seq = KinematicChainSeq(
             bounds_dof=BOUNDS,
             legs_list=["RF", "LF"],
             body_size=test_body_size,
         )
-        
+
         # Sequential processing
         seq_ik_sequential = LegInvKinSeq(
             aligned_pos=synthetic_pose_data,
             kinematic_chain_class=kinematic_chain_seq,
[…]
