### 2025-11-22 a7ea82e
fix more typos and inconsistencies in docs
