### 2025-11-22 18d59dc
minor doc/assertion updates
