### 2025-11-22 b75e389
Python 3.9 no more - EOL reached. This simplifies typehinting syntax
