+++ setup.py
@@ -4,7 +4,7 @@
 from setuptools import setup, find_packages
 
 setup(
-    python_requires=">3.8,<=3.14",
+    python_requires=">3.9,<=3.14",
     description="Inverse kinematics module for Drosophila",
     long_description=open("README.md").read(),
     long_description_content_type="text/markdown",
