### 2025-11-22 1ac28ea
Parallelize inverse and forward kinematics (#14)
* bump up python and numpy versions

* Parallelize at leg level

* add AI-generated tests

* wip - implement parallelization over time

* parallelize over time

* get parallelization example working and run scaling test on SCITAS

* refactor different processing modes into separate methods

* fix typos in parallel pipeline

* update unit tests for parallelization

* fix more typos and inconsistencies in docs

* minor doc/assertion updates

* Python 3.9 no more - EOL reached. This simplifies typehinting syntax
