+++ examples/example_leg_inv_kinematics_parallel.py
@@ -1,22 +1,19 @@
-"""
-    Example usage of leg inverse kinematics module.
-    It speeds up the process by running the pipeline in parallel.
-    Experiment on a Macbook Pro 2.3 GHz Quad-Core Intel Core i7, running IK on 6 legs:
-    Sequential IK took 1.7729304512341817 mins [serial]
-    Sequential IK took 0.58028298219045 mins [parallel]
+import matplotlib
 
-"""
-import time
-from pathlib import Path
+matplotlib.use("Agg")
 
+import pickle
+import joblib
 import numpy as np
 import matplotlib.pyplot as plt
-from multiprocessing import Pool
+from time import time
+from pathlib import Path
 
-from seqikpy.kinematic_chain import KinematicChainSeq, KinematicChainGeneric
-from seqikpy.leg_inverse_kinematics import LegInvKinSeq, LegInvKinGeneric
-from seqikpy.data import NMF_SIZE
+import seqikpy
+from seqikpy.kinematic_chain import KinematicChainSeq
+from seqikpy.leg_inverse_kinematics import LegInvKinSeq
 from seqikpy.utils import load_file, calculate_body_size
+
 
 TEMPLATE_NMF_LOCOMOTION = {
     "RF_Coxa": np.array([0.35, -0.27, 0.400]),
@@ -140,71 +137,171 @@ BOUNDS_LOCOMOTION = {
 }
 
 
-def worker_wrapper(aligned_pos, single_leg):
-    """ Wrapper for the function to run single leg IK. """
-    kin_chain = KinematicChainSeq(
-        bounds_dof=BOUNDS_LOCOMOTION,
-        body_size=calculate_body_size(
-            TEMPLATE_NMF_LOCOMOTION,
-            [single_leg]
-        ),
-        legs_list=[single_leg],
-    )
-    class_seq_ik = LegInvKinSeq(
-        aligned_pos=aligned_pos,
-        kinematic_chain_class=kin_chain,
-        initial_angles=INITIAL_ANGLES_LOCOMOTION,
-        log_level="ERROR"
-    )
-    leg_joint_angles, forward_kinematics = class_seq_ik.run_ik_and_fk(hide_progress_bar=True)
-    return leg_joint_angles, forward_kinematics
-
-
 if __name__ == "__main__":
-    # Aligned pose
-    data_path = Path("../data/df3d_pose_result__210902_PR_Fly1")
-    pose_data = load_file(
-        data_path / "pose3d_aligned.pkl"
+    # This script runs a weak scaling test (sequence length is proportional to number
+    # of workers. Change the multiplier below to global adjust sequence lengths.
+    task_size_multiplier = 1
+    hide_progress_bar = True
+
+    # Load aligned pose
+    data_dir = (
+        Path(seqikpy.__path__[0]).parent / "data/df3d_pose_result__210902_PR_Fly1"
+    )
+    aligned_pose_data_original = load_file(data_dir / "pose3d_aligned.pkl")
+
+    def _generate_input_sequence(n_repeats):
+        """Artificially repeat data to increase sequence length for benchmarking"""
+        repeated_pose_data = {}
+        for key, arr in aligned_pose_data_original.items():
+            arr_back = arr[::-1, ...]
+            repeated_pose_data[key] = np.concatenate(
+                [arr, arr_back] * (n_repeats // 2) + [arr] * (n_repeats % 2), axis=0
+            )
+        return repeated_pose_data
+
+    def _run_invik_pipeline(n_workers, **kwargs):
+        n_repeats = n_workers * task_size_multiplier
+        aligned_pose_data = _generate_input_sequence(n_repeats)
+        seq_length = aligned_pose_data[list(aligned_pose_data.keys())[0]].shape[0]
[…]
