### 2025-11-22 3cca896
Consolidate hard-coded constants into dataclass. Simplify logging convention in the meantime.
