### 2025-11-22 2787e21
introduce new body_config module manage constant deprecation
