+++ tests/test_parallelization.py
@@ -27,7 +27,7 @@ from pathlib import Path
 import seqikpy
 from seqikpy.leg_inverse_kinematics import LegInvKinSeq, LegInvKinGeneric
 from seqikpy.kinematic_chain import KinematicChainSeq, KinematicChainGeneric
-from seqikpy.data import neuromechfly_body_config
+from seqikpy.body_config import neuromechfly_body_config
 from seqikpy.utils import (
     calculate_body_size,
     split_arrays_into_chunks,
