+++ seqikpy/data.py
@@ -1,261 +1,24 @@
-"""Data, constants, and paths."""
+"""Deprecated data module."""
 
-import numpy as np
 import warnings
-from dataclasses import dataclass
-from copy import deepcopy as _deepcopy
+from .body_config import neuromechfly_body_config as _default_nmf_config
 
-
-@dataclass
-class BodyConfig:
-    # Size of the template body segments
-    segment_sizes: dict[str, float]
-
-    # Key points to align, to be provided in the alignment.Align class
-    points_to_align: dict[str, list[str]]
-
-    # Key points that are used in alignment
-    skeleton: list[str]
-
-    # Pose of each body landmark in the NeuroMechFly v0.0.6 model
-    # Note that each leg segment represents the joint in the proximal part
-    # For example, RF_Coxa means Thorax-Coxa joint
-    template: dict[str, np.ndarray]
-
-    # Initial joint angles for each leg and stage
-    initial_angles_rad: dict[str, dict[str, np.ndarray]] | None = None
-
-    # Lower bound of a DOF should be strictly lower than the initial angle.
-    # Upper bound of a DOF should be strictly bigger than the initial angle.
-    dof_bounds_rad: dict[str, tuple[float, float]] | None = None
-
-    def get_copy_of_initial_angles_in_deg(self):
-        initial_angles_deg = {}
-        for leg_name, stages in self.initial_angles_rad.items():
-            initial_angles_deg[leg_name] = {}
-            for stage, angles_rad in stages.items():
-                initial_angles_deg[leg_name][stage] = np.rad2deg(angles_rad)
-        return initial_angles_deg
-
-    def get_copy_of_dof_bounds_in_deg(self):
-        if self.dof_bounds_rad is None:
-            return None
-        dof_bounds_deg = {}
-        for dof, (lower_rad, upper_rad) in self.dof_bounds_rad.items():
-            dof_bounds_deg[dof] = (np.rad2deg(lower_rad), np.rad2deg(upper_rad))
-        return dof_bounds_deg
-
-    def set_initial_angles_in_deg(
-        self, initial_angles_deg: dict[str, dict[str, np.ndarray]]
-    ):
-        initial_angles_rad = {}
-        for leg_name, stages in initial_angles_deg.items():
-            initial_angles_rad[leg_name] = {}
-            for stage, angles_deg in stages.items():
-                initial_angles_rad[leg_name][stage] = np.deg2rad(angles_deg)
-        self.initial_angles_rad = initial_angles_rad
-
-    def set_dof_bounds_in_deg(self, dof_bounds_deg: dict[str, tuple[float, float]]):
-        dof_bounds_rad = {}
-        for dof, (lower_deg, upper_deg) in dof_bounds_deg.items():
-            dof_bounds_rad[dof] = (np.deg2rad(lower_deg), np.deg2rad(upper_deg))
-        self.dof_bounds_rad = dof_bounds_rad
-
-    def deepcopy(self):
-        return _deepcopy(self)
-
-
[…]
