### 2025-11-22 149917b
fix typos and filter out dof type warnings from ikpy
