### 2025-11-22 742a19a
Consolidate hard-coded constants into dataclass (#16)
* Consolidate hard-coded constants into dataclass. Simplify logging convention in the meantime.

* introduce new body_config module manage constant deprecation

* fix typos and filter out dof type warnings from ikpy
