### 2025-11-22 c044823
fix various doc site formatting issue and fix jupyter-book version to <2.
