+++ docs/requirements.txt
@@ -1,4 +1,4 @@
-jupyter-book
+jupyter-book<2
 matplotlib
 numpy
 ghp-import
