### 2025-11-22 bae7eaf
remove largely duplicated walking tutorial from main menu of the doc site
