### 2025-11-23 d7e705d
Add explicit instruction for cloning the directory if the data folder is needed
