### 2025-11-23 5cd6257
update python version in CI to match new python version range
