### 2025-11-23 539cc1a
typo/doc improvement
