### 2025-11-23 f92162d
Style/docs update (#17)
* run black

* update contributing.md

* fix various doc site formatting issue and fix jupyter-book version to <2.

* remove largely duplicated walking tutorial from main menu of the doc site

* Add explicit instruction for cloning the directory if the data folder is needed

* fix link to nmf zero pose image

* update python version in CI to match new python version range

* typo/doc improvement
