+++ docs/conf.py
@@ -1,9 +1,10 @@
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../seqikpy'))
+
+sys.path.insert(0, os.path.abspath("../seqikpy"))
 
 extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.napoleon',
-    'sphinx.ext.viewcode',
-    ]
+    "sphinx.ext.autodoc",
+    "sphinx.ext.napoleon",
+    "sphinx.ext.viewcode",
+]
