+++ examples/example_alignment.py
@@ -1,19 +1,18 @@
 """
     Examples showing how to use the alignment class in three different ways.
 """
-import pickle
 from pathlib import Path
 
 from seqikpy.alignment import AlignPose, convert_from_anipose_to_dict
 from seqikpy.body_config import neuromechfly_body_config
-from seqikpy.utils import load_file, save_file
+from seqikpy.utils import load_file
 
 
 if __name__ == "__main__":
-    # Case 1: we have pose3d, and we would like to align it but first
-    # we need to convert the pose data into a dictionary format
+    # Case 1: we have pose3d, and we would like to align it but first
+    # we need to convert the pose data into a dictionary format
 
-    DATA_PATH = Path('../data/anipose_220525_aJO_Fly001_001/pose-3d')
+    DATA_PATH = Path("../data/anipose_220525_aJO_Fly001_001/pose-3d")
 
     align = AlignPose.from_file_path(
         main_dir=DATA_PATH,
@@ -28,9 +27,9 @@ if __name__ == "__main__":
 
     aligned_pos = align.align_pose(export_path=DATA_PATH)
 
-    # Case 2: we have a dictionary format pose data, we just want to load and align it
+    # Case 2: we have a dictionary format pose data, we just want to load and align it
 
-    DATA_PATH = Path('../data/anipose_220525_aJO_Fly001_001/pose-3d')
+    DATA_PATH = Path("../data/anipose_220525_aJO_Fly001_001/pose-3d")
 
     align = AlignPose.from_file_path(
         main_dir=DATA_PATH,
@@ -44,10 +43,10 @@ if __name__ == "__main__":
 
     aligned_pos = align.align_pose(export_path=DATA_PATH)
 
-    # Case 3: we have the dictionary format pose data loaded already, we want to feed it
+    # Case 3: we have the dictionary format pose data loaded already, we want to feed it
     # to the class and align the pose. This assumes that the pose data is already aligned
 
-    DATA_PATH = Path('../data/anipose_220525_aJO_Fly001_001/pose-3d')
+    DATA_PATH = Path("../data/anipose_220525_aJO_Fly001_001/pose-3d")
 
     f_path = DATA_PATH / "converted_dict.pkl"
 
