+++ docs/methodology_alignment.md
@@ -10,7 +10,7 @@ The computation of joint angles could serve two primary purposes:
 
 We first map and align the experimentally acquired 3D poses with a biomechanical model's body [NeuroMechFly](https://github.com/NeLy-EPFL/NeuroMechFly). The alignment process consists of two stages. At the initial stage, we calculate the distances between specific body landmarks to derive the scaling constant to magnify or reduce the size of the experimental 3D data to match the body proportions of the biomechanical model (shown in the Figure below).
 
-```{image} ./images/NMF_zero_pose.png
+```{image} ./_images/NMF_zero_pose.png
 :width: 800px
 :align: center
 :name: biomech-model
