+++ docs/_toc.yml
@@ -17,10 +17,8 @@ parts:
 - caption: 📚 Tutorials
   chapters:
   - file: seqikpy_grooming
-  - file: seqikpy_locomotion
   - file: visualization
 - caption: 💬 References
   chapters:
-  - file: seqikpy.rst
-  - file: modules.rst
+  - file: seqikpy
   - file: references
