### 2025-11-23 fc49d4e
make typehint syntax compatible with python 3.9
