+++ setup.py
@@ -16,7 +16,7 @@ setup(
         "pandas",
         "joblib>=1.5,<2",
     ],
-    extras_require={"dev": ["pytest", "ImageHash", "black==22.3.0"]},
+    extras_require={"dev": ["pytest", "ImageHash", "black==25.11.0"]},
     license="Apache License 2.0",
     package_data={"seqikpy": ["data/*"]},
     include_package_data=True,
@@ -29,7 +29,12 @@ setup(
     classifiers=[
         "Development Status :: 4 - Beta",
         "Operating System :: OS Independent",
-        "Programming Language :: Python :: 3",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
+        "Programming Language :: Python :: 3.12",
+        "Programming Language :: Python :: 3.13",
+        "Programming Language :: Python :: 3.14",
         "License :: OSI Approved :: Apache Software License",
         "Topic :: Software Development :: Libraries",
     ],
