### 2025-11-23 0e8b895
Add "calibration" terminology to alignment documentation and comments
