+++ docs/paper/paper.md
@@ -43,9 +43,11 @@ Over the past decade, deep-learning based computer vision algorithms have transf
 
 Inverse Kinematics (IK) spans multiple domains including robotics, biomechanics, and character animation [@aristidou:2018]. In robotics, IK typically computes joint angles to achieve a desired end-effector position while respecting joint constraints. By contrast, in biomechanics, IK algorithms calculate joint angles to track all marker positions rather than only a single end-effector. This process is also known as multi-body kinematics optimization and is a well-established area in human biomechanics research [@delp:2007; @begon:2018; @pagnon:2022; @werling:2023].
 
-However, multi-body kinematics optimization is still an emerging field in insect research. Existing methods fall on two extremes along a spectrum of complexity. Simple approaches have been employed to compute joint angles using the dot product of two consecutive body segment vectors [@lobato-rios:2022;  @karashchuk:2021]. This often results in deviations from reference poses due to the lack of iterative corrections. On the other hand, more advanced gradient-based optimization methods have been developed to estimate joint angles in a biomechanical model of the fly [@vaxenburg:2024], using simulation-based Jacobians in physics-engines like MuJoCo [@todorov:2012]. Although these methods provide higher accuracy, they require complex setups that are not easily accessible to all researchers. Therefore, there is no stand-alone inverse kinematics software satisfying the trade-off between accuracy and complexity.
+However, multi-body kinematics optimization is still an emerging field in insect research. Existing methods fall on two extremes along a spectrum of complexity. Simple approaches have been employed to compute joint angles using the dot product of two consecutive body segment vectors [@lobato-rios:2022;  @karashchuk:2021]. This often results in deviations from reference poses due to the lack of iterative corrections. On the other hand, more advanced gradient-based optimization methods have been developed to estimate joint angles in a biomechanical model of the fly [@vaxenburg:2024], using simulation-based Jacobians in physics-engines like MuJoCo [@todorov:2012]. Although these methods provide higher accuracy, they are often entangled with pose estimation, rendering, or physics simulation frameworks (e.g. MuJoCo and various pose estimation postprocessing pipelines). This leads to complex dpendencies and heavy overhead. Here, we aim to provide lightweight and stand-alone package focused on inverse kinematics in a modular way.
 
 To address this gap, we have developed `SeqIKPy`, a fast and lightweight Python package for multi-body kinematics optimization in insects. Our package has two main stages: marker registration (aligning joint markers to a template in 3D) and inverse kinematics. The second stage draws upon the open-source IKPy [@Manceron_IKPy] library in a sequential manner. Through 3D visualizations, we demonstrate that our package can reliably reconstruct body kinematics for various fruit fly behaviors. Furthermore, recent studies have shown that the joint angles computed using `SeqIKPy` can accurately replicate animal behavior, both ground walking [@wang:2024] and grooming [@ozdil:2024], in physics-based simulations. Although our examples focus on the fly, our package's modular design allows for customization and extension to other animals with a similar body morphology.
+
[…]
