### 2025-11-23 45a964b
add background info on kalman filter and DL-based ik approaches. Recompile paper pdf
