+++ docs/paper/paper.md
@@ -35,29 +35,29 @@ Our framework contains:
 * Inverse kinematics: calculate joint angles sequentially from 3D poses.
 * Visualization: plot and animate the results in 3D.
 
-`SeqIKPy` is aimed at researchers studying detailed joint motion in animals with complex, multiple degrees-of-freedom body appendages. We provide examples for the fruit fly, *Drosophila melanogaster*. However, each module can easily be extended to be used with another model organism; the only requirements are the 3D kinematics of the target animal and its corresponding kinematic chain. Our package requires minimal Python knowledge and we provide extensive tutorials at [https://nely-epfl.github.io/sequential-inverse-kinematics](https://nely-epfl.github.io/sequential-inverse-kinematics).
+`SeqIKPy` is aimed at researchers studying detailed joint motion in animals with complex, multiple degrees-of-freedom body appendages. We provide examples for the fruit fly, *Drosophila melanogaster*. However, each module can easily be extended to be used with another model organism; the only requirements are the 3D kinematics of the target animal and its corresponding kinematic chain. Our package requires minimal Python knowledge, and extensive tutorials at [https://nely-epfl.github.io/sequential-inverse-kinematics](https://nely-epfl.github.io/sequential-inverse-kinematics) are available to novice users.
 
 # Statement of need
 
-Over the past decade, deep-learning based computer vision algorithms have transformed the analysis of behaviors in laboratory animals [@pereira:2020], including for the widely-used model organism, *Drosophila melanogaster*. Recently, researchers have developed deep learning-based 3D pose estimation tools [@gunel:2019; @karashchuk:2021] and detailed biomechanical models [@lobato-rios:2022; @wang:2024 ; @vaxenburg:2024], creating a growing need for tools to obtain more detailed descriptions of how body parts move in joint space. These computed joint angles can be replayed in physics-based simulations to estimate unmeasured physical quantities like joint torques [@lobato-rios:2022].
+Over the past decade, deep-learning based computer vision algorithms have transformed the analysis of behaviors in laboratory animals [@pereira:2020]. Recently, researchers have developed deep learning-based 3D pose estimation tools [@gunel:2019; @karashchuk:2021] and detailed biomechanical models [@lobato-rios:2022; @wang:2024 ; @vaxenburg:2024], creating a growing need for tools to obtain more detailed descriptions of how body parts move in joint-angle space. These computed joint angles can be replayed in physics-based simulations to estimate unmeasured physical quantities like joint torques [@lobato-rios:2022].
 
 Inverse Kinematics (IK) spans multiple domains including robotics, biomechanics, and character animation [@aristidou:2018]. In robotics, IK typically computes joint angles to achieve a desired end-effector position while respecting joint constraints. By contrast, in biomechanics, IK algorithms calculate joint angles to track all marker positions rather than only a single end-effector. This process is also known as multi-body kinematics optimization and is a well-established area in human biomechanics research [@delp:2007; @begon:2018; @pagnon:2022; @werling:2023].
 
[…]
