+++ docs/paper/paper.bib
@@ -182,7 +182,7 @@
   author = {Delcomyn,  Fred},
   year = {2004},
   month = jan,
-  pages = {51–70}
+  pages = {51-70}
 }
 
 @article{bates:2025,
@@ -193,4 +193,84 @@
   author = {Bates,  Alexander Shakeel and Phelps,  Jasper S and Kim,  Minsu and Yang,  Helen H and Matsliah,  Arie and Ajabi,  Zaki and Perlman,  Eric and Delgado,  Kevin M and Osman,  Mohammed Abdal Monium and Salmon,  Christopher K and Gager,  Jay and Silverman,  Benjamin and Renauld,  Sophia and Collie,  Matthew F and Fan,  Jingxuan and Pacheco,  Diego A and Zhao,  Yunzhi and Patel,  Janki and Zhang,  Wenyi and Serratosa Capdevilla,  Laia and Roberts,  Ruairi JV and Munnelly,  Eva J and Griggs,  Nina and Langley,  Helen and Moya-Llamas,  Borja and Maloney,  Ryan T and Yu,  Szi-chieh and Sterling,  Amy R and Sorek,  Marissa and Kruk,  Krzysztof and Serafetinidis,  Nikitas and Dhawan,  Serene and Stuerner,  Tomke and Klemm,  Finja and Brooks,  Paul and Lesser,  Ellen and Jones,  Jessica M and Pierce-Lundgren,  Sara E and Lee,  Su-Yee and Luo,  Yichen and Cook,  Andrew P and McKim,  Theresa H and Kophs,  Emily C and Falt,  Tjalda and Negron Morales,  Alexa M and Burke,  Austin and Hebditch,  James and Willie,  Kyle P and Willie,  Ryan and Popovych,  Sergiy and Kemnitz,  Nico and Ih,  Dodam and Lee,  Kisuk and Lu,  Ran and Halageri,  Akhilesh and Bae,  J. Alexander and Jourdan,  Ben and Schwartzman,  Gregory and Demarest,  Damian D and Behnke,  Emily and Bland,  Doug and Kristiansen,  Anne and Skelton,  Jaime and Stocks,  Tom and Garner,  Dustin and Salman,  Farzaan and Daly,  Kevin C and Hernandez,  Anthony and Kumar,  Sandeep and Dorkenwald,  Sven and Collman,  Forrest and Suver,  Marie P and Fenk,  Lisa M and Pankratz,  Michael J and Jefferis,  Gregory SXE and Eichler,  Katharina and Seeds,  Andrew M and Hampel,  Stefanie and Agrawal,  Sweta and Zandawala,  Meet and Macrina,  Thomas and Adjavon,  Diane-Yayra and Funke,  Jan and Tuthill,  John C and Azevedo,  Anthony and Seung,  H. Sebastian and de Bivort,  Benjamin L and Murthy,  Mala and Drugowitsch,  Jan and Wilson,  Rachel I and Lee,  Wei-Chung Allen},
   year = {2025},
   month = aug 
+}
+
+@software{aversiveplusplus_ik,
+  author    = {Dauphin, Loïc},
+  title     = {{ik}},
+  year      = {2016},
+  publisher = {GitHub},
+  journal   = {GitHub repository},
+  url       = {https://github.com/AversivePlusPlus/ik}
+}
+
+@article{Fohanno2010,
+  title = {Estimation of the 3D kinematics in kayak using an extended Kalman filter algorithm: a pilot study},
+  volume = {13},
+  ISSN = {1476-8259},
+  url = {http://dx.doi.org/10.1080/10255842.2010.491958},
+  DOI = {10.1080/10255842.2010.491958},
+  number = {sup1},
+  journal = {Computer Methods in Biomechanics and Biomedical Engineering},
[…]
