+++ docs/paper/README.md
@@ -1,0 +1,10 @@
+Compile PDF using a docker container:
+
+```bash
+# cd into this directory
+docker run --rm \
+    --volume $PWD/paper:/data \
+    --user $(id -u):$(id -g) \
+    --env JOURNAL=joss \
+    openjournals/inara
+```
