### 2025-11-23 0c17541
keep pdf in repo after all - it's managed by ci
