### 2025-11-23 d8e9391
add new tutorial on parallel processing
