### 2025-11-23 e1477c9
add information on optimization of all dofs instead of sequentially
