### 2025-11-23 a2309c2
Revision - paper text (#18)
* update paper

* (auto) Paper PDF Draft

* add background info on kalman filter and DL-based ik approaches. Recompile paper pdf

* (auto) Paper PDF Draft

* fix typo

* keep pdf in repo after all - it's managed by ci

* (auto) Paper PDF Draft

* add information on optimization of all dofs instead of sequentially

* (auto) Paper PDF Draft

---------

Co-authored-by: sibocw <sibocw@users.noreply.github.com>
