+++ README.md
@@ -85,7 +85,7 @@ If you find this package useful in your research, please consider citing it usin
 
 ```bibtex
 @software{ozdil2024seqikpy,
-  author       = {Ozdil, Pembe Gizem and Ijspeert, Auke and Ramdya, Pavan},
+  author       = {{\"O}zdil, Pembe Gizem and Wang-Chen, Sibo and Ning, Chuanfang and Ijspeert, Auke and Ramdya, Pavan},
   title        = {sequential-inverse-kinematics: v1.0.0},
   month        = jun,
   year         = 2024,
