### 2025-11-23 3f92ca2
add ack info for sibo
