### 2025-12-11 12ce475
[DOCS] fix mistakes in some files, address reviewer comments
