### 2025-12-11 c40644b
[EXP] add data and code for mouse reaching IK
