+++ examples/example_mouse_reaching.py
@@ -1,0 +1,469 @@
+"""
+Sequential inverse kinematics for a mouse forearm using SeqIKPy.
+
+We model a single arm with:
+- Shoulder yaw
+- Shoulder pitch
+- Shoulder roll
+- Elbow pitch
+
+Two IK stages:
+1) Use elbow trajectory to get shoulder yaw + pitch.
+2) Use wrist trajectory to get shoulder roll + elbow pitch (with yaw/pitch fixed).
+
+Input:
+    arm_markers: np.ndarray of shape (n_frames, 3, 3)
+        arm_markers[t, 0] = shoulder position at time t (3D)
+        arm_markers[t, 1] = elbow position at time t (3D)
+        arm_markers[t, 2] = wrist position at time t (3D)
+"""
+
+from dataclasses import dataclass
+from typing import Dict, Tuple, Optional
+from pathlib import Path
+
+import numpy as np
+import pandas as pd
+import matplotlib.pyplot as plt
+from tqdm import trange
+from ikpy.chain import Chain
+from ikpy.link import OriginLink, URDFLink
+
+from seqikpy.leg_inverse_kinematics import LegInvKinBase
+from seqikpy.visualization import plot_3d_points, animate_3d_points
+
+
+# Simple axes helpers
+X_AXIS = [1.0, 0.0, 0.0]
+Y_AXIS = [0.0, 1.0, 0.0]
+Z_AXIS = [0.0, 0.0, 1.0]
+
+
+@dataclass
+class MouseArmSegmentLengths:
+    upper_arm: float  # shoulder -> elbow
+    forearm: float  # elbow -> wrist
+
+
+class MouseArmIKSeq(LegInvKinBase):
+    """
+    Sequential inverse kinematics for a single mouse arm.
+
+    Stages:
+    - Stage 1: shoulder yaw + pitch (2 DOF), using the elbow as end-effector.
+    - Stage 2: shoulder roll + elbow pitch (2 DOF), using the wrist as end-effector
+               while keeping yaw & pitch fixed from Stage 1.
+    """
+
+    def __init__(
+        self,
+        arm_markers: np.ndarray,
+        segment_lengths: Optional[MouseArmSegmentLengths] = None,
+    ) -> None:
+        if (
+            arm_markers.ndim != 3
+            or arm_markers.shape[2] != 3
+        ):
+            raise ValueError(
+                "arm_markers must have shape (n_frames, 3, 3): "
+                "[shoulder, elbow, wrist] x [x,y,z]"
+            )
+
+        self.arm_markers = arm_markers
+        self.n_frames = arm_markers.shape[0]
+
+        if segment_lengths is None:
+            self.segment_lengths = self._estimate_segment_lengths(arm_markers)
+        else:
+            self.segment_lengths = segment_lengths
+
+        self.joint_angles_dict: Dict[str, np.ndarray] = {}
+        self.forward_kinematics: Dict[str, np.ndarray] = {}
+
+    @staticmethod
+    def _estimate_segment_lengths(arm_markers: np.ndarray) -> MouseArmSegmentLengths:
+        """Estimate segment lengths (upper arm, forearm) as median distances."""
+        shoulder = arm_markers[:, 0, :]
+        elbow = arm_markers[:, 1, :]
+        wrist = arm_markers[:, 2, :]
+
+        upper_arm_lengths = np.linalg.norm(elbow - shoulder, axis=1)
+        forearm_lengths = np.linalg.norm(wrist - elbow, axis=1)
+
+        upper_arm = np.median(upper_arm_lengths)
+        forearm = np.median(forearm_lengths)
+
+        return MouseArmSegmentLengths(
+            upper_arm=float(upper_arm),
[…]
