### 2025-12-11 96a9f21
Minor changes; remove unused functions
