### 2025-12-11 6b9b460
Minor changes: use uv instead of pip; matplotlib version set; update gitignore
