+++ .github/workflows/CI.yml
@@ -17,7 +17,7 @@ jobs:
 
     strategy:
       matrix:
-        python-version: ["3.9", "3.14"]  # oldest and newest supported versions
+        python-version: ["3.10", "3.14"]  # oldest and newest supported versions
 
     steps:
       - name: Checkout repository
