### 2025-12-12 eda5ff8
[PAPER] address reviewer comments; minor edits
