+++ docs/paper/paper.md
@@ -42,13 +42,13 @@ Our framework contains:
 
 # Statement of need
 
-Over the past decade, deep-learning based computer vision algorithms have transformed the analysis of behaviors in laboratory animals [@pereira:2020]. Recently, researchers have developed deep learning-based 3D pose estimation tools [@gunel:2019; @karashchuk:2021] and detailed biomechanical models [@lobato-rios:2022; @wang:2024 ; @vaxenburg:2024], creating a growing need for tools to obtain more detailed descriptions of how body parts move in joint-angle space. These computed joint angles can be replayed in physics-based simulations to estimate unmeasured physical quantities like joint torques [@lobato-rios:2022].
+Over the past decade, deep-learning–based computer vision algorithms have transformed behavioral analysis in laboratory animals [@pereira:2020]. More recently, deep-learning–based 3D pose estimation tools [@gunel:2019; @karashchuk:2021] and detailed biomechanical models [@lobato-rios:2022; @wang:2024; @vaxenburg:2024] have created a growing need for tools that translate 3D kinematics into joint-angle representations. These joint angles can then be replayed in physics-based simulations to estimate unmeasured physical quantities such as joint torques [@lobato-rios:2022].
 
-Inverse Kinematics (IK) spans multiple domains including robotics, biomechanics, and character animation [@aristidou:2018]. In robotics, IK typically computes joint angles to achieve a desired end-effector position while respecting joint constraints. By contrast, in biomechanics, IK algorithms calculate joint angles to track all marker positions rather than only a single end-effector. This process is also known as multi-body kinematics optimization and is a well-established area in human biomechanics research [@delp:2007; @begon:2018; @pagnon:2022; @werling:2023].
+Inverse kinematics (IK) is widely used in robotics, biomechanics, and character animation [@aristidou:2018]. In robotics, IK typically computes joint angles to achieve a desired end-effector position while respecting joint constraints of a robot. In contrast, biomechanical IK aims to track multiple body markers simultaneously, a process often referred to as multi-body kinematics optimization and well established in human biomechanics [@delp:2007; @begon:2018; @pagnon:2022; @werling:2023].
 
-However, multi-body kinematics optimization is still an emerging field in insect research. Existing methods fall on two extremes along a spectrum of complexity. Simple approaches have been employed to compute joint angles using the dot product of two consecutive body segment vectors [@lobato-rios:2022;  @karashchuk:2021]. This often results in deviations from reference poses due to the lack of iterative corrections. On the other hand, more advanced gradient-based optimization methods have been developed to estimate joint angles in a biomechanical model of the fly [@vaxenburg:2024], using simulation-based Jacobians in physics-engines like MuJoCo [@todorov:2012]. Although these methods provide higher accuracy, they are often entangled with pose estimation, rendering, or physics simulation frameworks (e.g. MuJoCo and various pose estimation postprocessing pipelines). This leads to complex dependencies and heavy overhead. Here, we aim to provide a lightweight and stand-alone package focused on inverse kinematics in a modular way.
[…]
