### 2025-12-12 fb96b07
[FIX] hopefully fix the failing test in python 3.14
