### 2025-12-12 89c784e
[DOCS] add mouse example to the website, minor edits to the text
