+++ docs/paper/paper.md
@@ -50,7 +50,8 @@ In insect research, however, multi-body kinematics optimization remains relative
 
 `SeqIKPy` consists of two main stages: marker registration (aligning measured keypoints to a 3D body template) and inverse kinematics. The latter stage builds on the open-source IKPy library [@Manceron_IKPy] and applies it sequentially along the kinematic chain. Using 3D visualizations, we show that SeqIKPy reliably reconstructs fly kinematics across a range of behaviors. Previous work has demonstrated that the joint angles computed with SeqIKPy accurately reproduce both walking [@wang:2024] and grooming [@ozdil:2024] behaviors in physics-based simulations. Although our examples mostly focus on the fly, we demonstrate that its modular design allows adaptation to other animals with articulated limbs, including a mouse forelimb.
 
-`SeqIKPy` can be used for animals and robots with arbitrarily configured kinematic chains consisting of rigid bodies connected with rotational joints. However, we have focused on the fruit fly _Drosophila melanogaster_ in our demonstrations. Insects are some of the oldest model organisms in the study of motor control [@delcomyn:2004]. In particular, the fly _Drosophila melanogaster_ is one of the most common model organisms in neuroscience, thanks to its compact but versatile nervous system. In the past few years, _Drosophila melanogaster_ has become the most complex organism whose entire central nervous system has been mapped out [for example, see @bates:2025]. This resulted in a rapidly growing community of researchers in the field and a renewed demand for open-source data processing tools. With `SeqIKPy`, we aim to improve a crucial step in the data processing pipeline for behavior analysis.
+`SeqIKPy` can be used for animals and robots with arbitrarily configured kinematic chains consisting of rigid bodies connected with rotational joints. However, we have focused on the fruit fly _Drosophila melanogaster_ in our demonstrations. Insects are some of the oldest model organisms in the study of motor control [@delcomyn:2004], and _Drosophila melanogaster_ is particularly prominent in neuroscience research due to its compact yet versatile nervous system. Recent advances have made the fly the most complex organism with a fully mapped central nervous system [e.g., @bates:2025], leading to rapid growth in the field and an increased demand for open-source data-processing tools. With `SeqIKPy`, we aim to address a critical gap in the behavioral analysis pipeline.
+
 
 # Overview
 
