### 2025-12-12 #8557 @gizemozd
We have addressed all of the reviewer’s requests. The detailed code changes and text modifications can be found in [this PR](https://github.com/NeLy-EPFL/sequential-inverse-kinematics/pull/15).

Dear @aceglia, thank you very much for your thorough review, and we apologize for the delay in our response.

Below, we provide our point-by-point replies:

1. We have added an explanation of why it’s important to estimate kinematics for insects in the Statement of Need section.
2. SeqIKPy is not limited to insects or fly-specific anatomy. It supports any kinematic chain composed of rigid segments connected by rotational joints, regardless of species. By “similar body morphology,” we mean body parts that can be modeled as articulated chains with well-defined joint axes and segment lengths. The package was originally named for insects because the core classes use insect-specific leg terminology (e.g., Coxa, Femur). In practice, the framework can be extended to vertebrate limbs. To make this explicit, we have clarified the text and added an example showing how the base LegInvKin class can be adapted to compute joint angles for a mouse forelimb (see the related [tutorial](https://github.com/NeLy-EPFL/sequential-inverse-kinematics/pull/15/commits/89c784e7865b6d226129f078dc63d1d904b214ce)). This demonstrates that the same approach can be used, with modest modifications, for quadrupedal or bipedal animals.
3. We thank the reviewer for this suggestion. We have added (1) guideline for formatting with Black (compliance is validated through CI/CD), (2) guideline for using numpydoc for documentation, (3) guideline for using PyTest for documentation, and (4) information about the core abstract base classes in the Get Started section of the contribution guide.
4. We acknowledge that seqikpy (and indeed any inverse kinematics implementation based on per-frame optimization) are computation intensive. To mitigate this, we have now parallelized the inverse and forward kinematics computation (over different kinematic chains and over time). Each kinematic time series is split into several chunks. Chunk sizes are adaptively chosen to balance the tradeoff between load balancing and overhead. Adjacent chunks share a small overlap in order to mitigate boundary effect due to re-initialization at the beginning of each atomic task.  Accordingly, we have added a new [tutorial](https://github.com/NeLy-EPFL/sequential-inverse-kinematics/pull/19) on parallel processing.. We conducted a Weak Scaling Test (i.e. task size is proportional to the number of worker processes) and observed that throughput scales almost linearly with respect to the number of workers up to the number of physical CPU cores (and to a lesser extent, the number of logical CPU threads). Details can be found here: https://github.com/NeLy-EPFL/sequential-inverse-kinematics/pull/14. For example, on a 36-core (72-thread) Intel Xeon Platinum 8360Y processor, we obtained a ~32x speedup using 36 processes (90% efficiency), and a ~58x speedup using 72 processes (~80% efficiency). The overall throughput is ~2ms/frame/kinematic chain, which translates to $0.30 per hour of recording at 30 FPS with 6 kinematic chains. We have made references to alternative approaches using Kalman filters in the paper. Furthermore, we added a mention to Aversive++’s IK module https://github.com/AversivePlusPlus/ik in the paper. This module is similar to ikpy (in fact, it is referenced from the README of ikpy), but written in C++. This library is an interesting alternatives for users who do not require sequential optimization along the kinematic chain but require higher throughput.

5. We thank the reviewer for this suggestion. We have now consolidated constants related to specific body morphologies into a dataclass, and provided defaults for the fruit fly as an instance of this dataclass.
