### 2025-12-12 #8557 @gizemozd
We have addressed all of the reviewer’s requests. The detailed code changes and text modifications can be found in [this PR](https://github.com/NeLy-EPFL/sequential-inverse-kinematics/pull/15).

Dear @davidpagnon, thank you very much for your thorough review, and we apologize for the delay in our response.

Below, we provide our point-by-point replies:

### Paper
* We are referring to the fact that libraries implementing mature inverse kinematics computation are often very heavy and entangled with pose estimation, graphic rendering libraries, and even accelerated hardware (e.g. MuJoCo and various postprocessing pipelines for pose estimation frameworks). Here we aim to provide a lightweight and standalone implementation. We have clarified this in the text.
* We have included more discussion on this topic. Regarding processing time: we have implemented new parallel processing options with near-linear scalability up to hardware constraints. On a typical processor found in institution computing clusters (Intel Xeon Platinum 8360Y), we observed a final throughput of 2ms/frame/kinematic chain, which translates to ~$0.30 per hour of recording at 30 FPS with 6 kinematic chains—likely sufficient for most applications.

### Documentation and Tutorials
* We have clarified the term "alignment" in the text and documentation, but we decided to not change it in the code for the sake of API consistency.
* Thank you for pointing this out. We have now fixed the issue related to API Reference Section.
* We have now added a section to explain how obtain 3D pose estimation files from scratch.
* Thank you for pointing this out. We have fixed the link to the biomechanical model image.
* Thank you for pointing this out. We have fixed the data path in the grooming tutorial.
* The two tutorials indeed share a lot of information. We have removed the tutorial on locomotion, but retained a note referring to it from the grooming tutorial.
* Thank you for pointing this out. We opted to not include the data folder as package data to avoid bloating the package (this could, for example, slow down CI/CD workflows for packages requiring seqikpy). We have added clear instructions in the “Installation” section requesting the user to clone the repository instead of downloading from PyPI if the user intends to follow the tutorials or help develop the package.
* We have included a “hide_progress_bar” argument to the most time-consuming functions, controlling whether a progress bar would be shown.
* `TEMPLATE_NMF` is a dictionary of 3D joint locations in a defined reference (“zero/rest”) pose of the NeuromechFly model. We use it (1) to compute segment directions/lengths for pose alignment and (2) to define the fixed base/origin of each kinematic chain. The grooming vs locomotion templates differ because they were extracted from model configurations (i.e., leg positions where slightly changed), so the absolute joint coordinates (e.g., RF_Coxa) are different. We obtained these coordinates by reading joint locations from the 3D model in Blender. A point-and-click GUI would indeed be convenient, but Blender already provides a robust workflow for inspecting and exporting precise joint positions; therefore, we think this is outside of the scope of our package. Also, we provide **a built-in way (`[from_sdf](https://github.com/NeLy-EPFL/sequential-inverse-kinematics/blob/1598cf49a166f061695c01b5feace398be2a5cac/seqikpy/utils.py#L394 )`) to retrieve joint locations directly from the model definition (XML/SDF)**, so users don’t need to manually enter coordinates. We have made this feature explicit and highlight it in the paper.
* We have now consolidated constants related to specific body morphologies into a dataclass, and provided defaults for the fruit fly as an instance of this dataclass (see this [PR](https://github.com/NeLy-EPFL/sequential-inverse-kinematics/pull/16)).
* For consistency and compatibility with various other libraries that the user might use with seqikpy (e.g. physics simulation libraries), we’ve decided to keep angles represented in radians. However, for convenience, we have provided `get_copy_of_initial_angles_in_deg`, `get_copy_of_dof_bounds_in_deg`, `set_initial_angles_in_deg`, and `set_dof_bounds_in_deg` in the new `BodyConfig` dataclass. By design, we did not use the getter-and-setter design pattern (e.g. to have both `.initial_angles_rad` and `.initial_angles_deg` exposed as “attributes” while sharing the same private underlying data `._initial_angles_rad`). This is because we want to make it explicit that (1) the underlying (and therefore mutable) data structure stores data in radians, and (2) if the user prefers degrees, then the data in degrees are a copy of the underlying data instead of a view of them.
* Thank you for this thoughtful comment. You are right that the current package does not use a single unified kinematic chain. This limitation is not specific to the insect head but applies more generally to any body part that consists of multiple interdependent kinematic chains. In our case, the head comprises two coupled chains—neck–right antenna/eye and neck–left antenna/eye—rather than a single linear chain like the legs.
While such structures can, in principle, be represented within a unified inverse kinematics framework, this requires certain degrees of freedom to be shared across chains. In practice, this leads to compromises in the estimated joint angles depending on which chain is more strongly constrained by the tracked data. To avoid introducing these artifacts, we therefore implemented a dedicated head-kinematics dataset and class.
We will clarify this design choice and its implications more explicitly in the manuscript.
