### 2025-12-13 64c54f1
clarify python version requirement and clarify installation instructions (offer uv as an alternative to pip, but allow native pip as well)
