### 2025-12-13 f030e84
relax precision for floating point equality assertion in unit test
