### 2025-12-14 b602b62
improve body_config documentation and add warnings for unexpected values
