+++ seqikpy/body_config.py
@@ -1,55 +1,116 @@
+import logging
 import numpy as np
-from dataclasses import dataclass
 from copy import deepcopy as _deepcopy
 from typing import Optional
 
+_logger = logging.getLogger(__name__)
 
-@dataclass
+
 class BodyConfig:
-    """
-    Configuration for the body model used in physics simulation.
 
-    Attributes:
-        segment_sizes (dict[str, float]): Size of the template body segments.
-        skeleton (list[str]): Key points that are used in alignment process.
-        points_to_align (dict[str, list[str]]): Body key points to align,
-            to be provided in `AlignPose`. Each key is the name of a kinematic
-            chain, and each value is a list of body keypoint names. All body
-            keypoint name must be found in `skeleton`.
-        template (dict[str, np.ndarray]): A template for the specified skeleton,
-            i.e. the xyz position of all body keypoints.
-        initial_angles_rad (Optional[dict[str, dict[str, np.ndarray]]]): Initial
-            joint angles for each leg and stage, in radians. The keys of the
-            outer dict are leg names (e.g., "RF", "LF"), and the keys of the
-            inner dict are stage names (e.g., "stage_1", "stage_2"). The values
-            of the inner dict are numpy arrays of joint angles relevant to that
-            processing stage in radians.
-        dof_bounds_rad (Optional[dict[str, tuple[float, float]]]): Lower and
-            upper bounds for each degree of freedom (DOF), in radians. Each key
-            is a DOF name (e.g., "RF_ThC_roll"), and each value is a tuple of
-            (lower_bound, upper_bound).
-    """
-
-    # Size of the template body segments
     segment_sizes: dict[str, float]
-
-    # Key points to align, to be provided in the alignment.Align class
     points_to_align: dict[str, list[str]]
-
-    # Key points that are used in alignment
     skeleton: list[str]
-
-    # Pose of each body landmark in the NeuroMechFly v0.0.6 model
-    # Note that each leg segment represents the joint in the proximal part
-    # For example, RF_Coxa means Thorax-Coxa joint
     template: dict[str, np.ndarray]
+    _initial_angles_rad: Optional[dict[str, dict[str, np.ndarray]]]
+    _dof_bounds_rad: Optional[dict[str, tuple[float, float]]]
 
-    # Initial joint angles for each leg and stage
-    initial_angles_rad: Optional[dict[str, dict[str, np.ndarray]]] = None
+    def __init__(
+        self,
+        segment_sizes: dict[str, float],
+        points_to_align: dict[str, list[str]],
+        skeleton: list[str],
+        template: dict[str, np.ndarray],
+        initial_angles_rad: Optional[dict[str, dict[str, np.ndarray]]] = None,
+        initial_angles_deg: Optional[dict[str, dict[str, np.ndarray]]] = None,
+        dof_bounds_rad: Optional[dict[str, tuple[float, float]]] = None,
+        dof_bounds_deg: Optional[dict[str, tuple[float, float]]] = None,
+    ):
+        """
+        Configuration for the body model used in physics simulation.
 
-    # Lower bound of a DOF should be strictly lower than the initial angle.
-    # Upper bound of a DOF should be strictly bigger than the initial angle.
-    dof_bounds_rad: Optional[dict[str, tuple[float, float]]] = None
+        Attributes:
+            segment_sizes (dict[str, float]): Size of the template body segments.
+            skeleton (list[str]): Key points that are used in alignment process.
+            points_to_align (dict[str, list[str]]): Body key points to align,
+                to be provided in `AlignPose`. Each key is the name of a kinematic
[…]
