### 2025-12-14 9ec5376
replace setup.py with pyproject.toml per PEP 621
