+++ setup.py
@@ -1,42 +1,0 @@
-#!/usr/bin/env python
-
-from setuptools import setup, find_packages
-
-setup(
-    python_requires=">3.9,<3.15",
-    description="Inverse kinematics module for Drosophila",
-    long_description=open("README.md").read(),
-    long_description_content_type="text/markdown",
-    install_requires=[
-        "ikpy>=3.3.4",
-        "opencv-python>4.5",
-        "numpy>=2.0,<3",
-        "tqdm",
-        "matplotlib",
-        "pandas",
-        "joblib>=1.5,<2",
-    ],
-    extras_require={"dev": ["pytest", "ImageHash", "black==25.11.0"]},
-    license="Apache License 2.0",
-    package_data={"seqikpy": ["data/*"]},
-    include_package_data=True,
-    name="seqikpy",
-    packages=find_packages(include=["seqikpy", "seqikpy.*"]),
-    test_suite="tests",
-    url="https://github.com/NeLy-EPFL/sequential-inverse-kinematics.git",
-    version="1.0.3",
-    zip_safe=False,
-    classifiers=[
-        "Development Status :: 4 - Beta",
-        "Operating System :: OS Independent",
-        "Programming Language :: Python :: 3.9",
-        "Programming Language :: Python :: 3.10",
-        "Programming Language :: Python :: 3.11",
-        "Programming Language :: Python :: 3.12",
-        "Programming Language :: Python :: 3.13",
-        "Programming Language :: Python :: 3.14",
-        "License :: OSI Approved :: Apache Software License",
-        "Topic :: Software Development :: Libraries",
-    ],
-    keywords="inverse kinematics, robotics, insects, motion analysis",
-)
