### 2025-12-15 a0adc58
[PAPER] add a sentence about sequential approach
