+++ docs/paper/paper.md
@@ -48,7 +48,7 @@ Inverse kinematics (IK) is widely used in robotics, biomechanics, and character 
 
 In insect research, however, multi-body kinematics optimization remains relatively underdeveloped. Existing approaches lie at two extremes. Simple geometric methods compute joint angles from dot products between adjacent body segments [@lobato-rios:2022; @karashchuk:2021], often leading to cumulative errors due to the lack of iterative corrections. More advanced gradient-based optimization methods estimate joint angles in a biomechanical model of the fly [@vaxenburg:2024], using simulation-based Jacobians in physics-engines like MuJoCo [@todorov:2012]. Although these methods provide higher accuracy, they are often entangled with pose estimation and physics engines, leading to complex dependencies and heavy overhead. Here, we introduce `SeqIKPy` as a lightweight, stand-alone, and modular solution focused on inverse kinematics.
 
-`SeqIKPy` consists of two main stages: marker registration (aligning measured keypoints to a 3D body template) and inverse kinematics. The latter stage builds on the open-source IKPy library [@Manceron_IKPy] and applies it sequentially along the kinematic chain. Using 3D visualizations, we show that SeqIKPy reliably reconstructs fly kinematics across a range of behaviors. Previous work has demonstrated that the joint angles computed with SeqIKPy accurately reproduce both walking [@wang:2024] and grooming [@ozdil:2024] behaviors in physics-based simulations. Although our examples mostly focus on the fly, we demonstrate that its modular design allows adaptation to other animals with articulated limbs, including a mouse forelimb.
+`SeqIKPy` consists of two main stages: marker registration (aligning measured keypoints to a 3D body template) and inverse kinematics. The latter stage builds on the open-source IKPy library [@Manceron_IKPy] and applies it sequentially along the kinematic chain. The sequential nature of our method constrains the solution locally and mitigates ambiguities arising from joint redundancy. Using 3D visualizations, we show that SeqIKPy reliably reconstructs fly kinematics across a range of behaviors. Previous work has demonstrated that the joint angles computed with SeqIKPy accurately reproduce both walking [@wang:2024] and grooming [@ozdil:2024] behaviors in physics-based simulations. Although our examples mostly focus on the fly, we demonstrate that its modular design allows adaptation to other animals with articulated limbs, including a mouse forelimb.
 
 `SeqIKPy` can be used for animals and robots with arbitrarily configured kinematic chains consisting of rigid bodies connected with rotational joints. However, we have focused on the fruit fly _Drosophila melanogaster_ in our demonstrations. Insects are some of the oldest model organisms in the study of motor control [@delcomyn:2004], and _Drosophila melanogaster_ is particularly prominent in neuroscience research due to its compact yet versatile nervous system. Recent advances have made the fly the most complex organism with a fully mapped central nervous system [e.g., @bates:2025], leading to rapid growth in the field and an increased demand for open-source data-processing tools. With `SeqIKPy`, we aim to address a critical gap in the behavioral analysis pipeline.
 
