### 2025-12-16 #8557 @olexandr-konovalov
That's all nice to read! @davidpagnon and @aceglia when the reviewers are satisfied with the improvements, we ask that they confirm their recommendation to accept the submission. If you are happy with that, could you please confirm that using the "accept"-like wording? Thanks!
