+++ docs/paper/paper.md
@@ -30,23 +30,23 @@ bibliography: paper.bib
 ---
 
 # Summary
-`SeqIKPy` is a Python package for inverse kinematics (IK) calculation in animal bodies with complex joint configurations. The name stands for Sequential Inverse Kinematics in Python, as our method computes joint angles sequentially by performing IK for each joint along a kinematic chain.
+`SeqIKPy` is a Python package for inverse kinematics (IK) calculations in animal bodies with complex joint configurations. The name stands for Sequential Inverse Kinematics in Python, as our method computes joint angles sequentially by performing IK for each joint along a kinematic chain.
 
 Our framework contains:
 
-* Pose alignment: map tracked key point locations in 3D onto an animal body template.
+* Pose alignment: map tracked keypoint locations in 3D onto an animal body template.
 * Inverse kinematics: calculate joint angles sequentially from 3D poses.
 * Visualization: plot and animate the results in 3D.
 
-`SeqIKPy` is aimed at researchers studying detailed joint motion in animals with complex, multiple degrees-of-freedom body appendages. We provide examples for the fruit fly, *Drosophila melanogaster*. However, each module can easily be extended to be used with another model organism; the only requirements are the 3D kinematics of the target animal and its corresponding kinematic chain. Our package requires minimal Python knowledge, and extensive tutorials are available to novice users at [https://nely-epfl.github.io/sequential-inverse-kinematics](https://nely-epfl.github.io/sequential-inverse-kinematics).
+`SeqIKPy` is aimed at researchers studying detailed joint motion in animals with complex, multiple degrees of freedom body appendages. We provide examples for the fruit fly, *Drosophila melanogaster*. However, each module can easily be extended to be used with other model organisms; the only requirements are the 3D kinematics of the target animal and its corresponding kinematic chain. Our package requires minimal Python knowledge, and extensive tutorials are available to novice users at [https://nely-epfl.github.io/sequential-inverse-kinematics](https://nely-epfl.github.io/sequential-inverse-kinematics).
 
 # Statement of need
 
 Over the past decade, deep-learning–based computer vision algorithms have transformed behavioral analysis in laboratory animals [@pereira:2020]. More recently, deep-learning–based 3D pose estimation tools [@gunel:2019; @karashchuk:2021] and detailed biomechanical models [@lobato-rios:2022; @wang:2024; @vaxenburg:2024] have created a growing need for tools that translate 3D kinematics into joint-angle representations. These joint angles can then be replayed in physics-based simulations to estimate unmeasured physical quantities such as joint torques [@lobato-rios:2022].
 
-Inverse kinematics (IK) is widely used in robotics, biomechanics, and character animation [@aristidou:2018]. In robotics, IK typically computes joint angles to achieve a desired end-effector position while respecting joint constraints of a robot. In contrast, biomechanical IK aims to track multiple body markers simultaneously, a process often referred to as multi-body kinematics optimization and well established in human biomechanics [@delp:2007; @begon:2018; @pagnon:2022; @werling:2023].
+Inverse kinematics (IK) is widely used in robotics, biomechanics, and character animation [@aristidou:2018]. In robotics, IK typically computes joint angles to achieve a desired end-effector position while respecting joint constraints of a robot. In contrast, biomechanical IK aims to track multiple body markers simultaneously, a process often referred to as multi-body kinematics optimization and well-established in human biomechanics [@delp:2007; @begon:2018; @pagnon:2022; @werling:2023].
 
[…]
