### 2025-12-21 8fdbd68
[CI] fix deploy-pages error
