### 2026-01-19 #8557 @olexandr-konovalov
@gizemozd thank you for the updates. I have noticed that https://doi.org/10.5281/zenodo.18010528 has the title "sequential-inverse-kinematics: v1.1.0" - please edit its description on Zenodo to have the title exactly as the paper: "SeqIKPy: a Python package for inverse kinematics in insects".

Furthermore, I've proofread the paper:

- p.3, line 111: use monospace font for `SeqIKPy` like elsewhere, for consistency

- p.4, lines 130, 146: fix capitalisation in "Kalman"

- p.4, line 158 (IKPy citation): compare with "Manceron, P. (2024). IKPy (v3.4.2). Zenodo. https://doi.org/10.5281/zenodo.13304536". First, specifying version is a good practice; second, if your software relies on it, perhaps you'd like to refer to the latest release

- p.4, line 137 (ik citation): unlike IKPy, there seem to be no releases, Zenodo archival etc., and even no license in this repository. However, it seems that `ik` is included into https://github.com/AversivePlusPlus/AversivePlusPlus/tree/v17.02-tag/tools/ik where it is a part of v17.02 release of Aversive++, with the latter repository having a proper BSD-3-Clause license. Wouldn't it be more appropriate to cite Aversive++ (all commits in the `v17.02-tag` branch belong the same and only author). For example, you can use something like
```
@misc{aversive,
  title = {Aversive++},
  author = {Loïc Dauphin},
  note = {https://aversiveplusplus.github.io/},
  version = {v17.02},
  year = {2017}
}
```
