### 2026-01-20 #8557 @sibocw
@olexandr-konovalov Thanks for proofreading the paper. I implemented the changes you suggested. For BibTeX entries for `ik` (and also `ikpy`) I used the following canonical format which seems to be rendered nicely:

```
@software{aversiveplusplus_ik,
  author    = {Dauphin, Loïc},
  license   = {BSD-3-Clause},
  title     = {{Aversive++}},
  version   = {v17.02},
  year      = {2017},
  publisher = {GitHub},
  url       = {https://github.com/AversivePlusPlus/AversivePlusPlus}
}
```

@gizemozd can you update the Zenodo record? I don't have write access to it.

Thanks!
