### 2026-01-20 #8557 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
