### 2026-01-20 #8557 @danielskatz
👋 @gizemozd - I've suggested a few minor changes in https://github.com/NeLy-EPFL/sequential-inverse-kinematics/pull/21 - please merge this, or let me know what disagree with
