### 2026-01-21 #8557 @gizemozd
> 👋 @gizemozd - I've suggested a few minor changes in https://github.com/NeLy-EPFL/sequential-inverse-kinematics/pull/21 - please merge this, or let me know what disagree with

@danielskatz merged! Thank you for the edits!
