### 2026-01-21 #8557 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
