### 2026-01-21 #8557 @danielskatz
Congratulations to @gizemozd (Pembe Gizem Ozdil) and co-authors on your JOSS publication!!

JOSS completely depends on volunteers, and we couldn't do this without you
